backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

sqlmodel==0.0.21
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.13.3

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

arq==0.26.1
redis==5.1.1

openai==1.51.0

pypdf==4.3.1
pdfplumber>=0.11
python-docx==1.1.2
aiofiles==24.1.0

pydantic-settings==2.5.2
pydantic[email]==2.9.2

cryptography==43.0.1
httpx==0.27.2

pytest==8.3.3
pytest-asyncio==0.24.0
